fix: resolve CI failures for Hytte-e313
Move complianceIcon/complianceBadgeClass/flagIsSevere out of DayCard.tsx
into strideHelpers.tsx so DayCard only exports components (fixes
react-refresh/only-export-components). Remove synchronous setState calls
in the !planId early-return of WeekDetailsModal useEffect (fixes
react-hooks/set-state-in-effect); the reset is unnecessary since the
component already returns null when week is null.
